requirements.txt
opencv-python
numpy
Pillow
viam-sdk==0.47.0

typing-extensions
